Remove unnecessary warnings.
* wflow_start() sent a warning about an existing Git repo even when the
user explicitly set existing=TRUE. This is an artifact. Now that the
issue of an existing Git repo is caught very early in the function, this
is no longer necessary.

* wflow_paths() sent a warning if the website directory didn't exist.
This doesn't actually hurt anything. Running wflow_build() or
wflow_publish() will create the directory.

* wflow_publish() sent a warning if the website directory didn't exist
because it tried unsuccessfully to create a backup copy. Now
wflow_build() does not attempt to copy if it doesn't exist (also made
the improvement that it doesn't copy if it's a dry run).

+++ R/wflow_publish.R
@@ -204,17 +204,19 @@ wflow_publish <- function(
   if (length(files_to_build) > 0) {
     # Create a backup copy of the docs/ directory. If either step 2 (build the
     # HTML) or step 3 (commit the HTML) fails, delete docs/ and restore backup
-    docs_backup <- tempfile(pattern = sprintf("docs-backup-%s-",
-                                              format(Sys.time(),
-                                                     "%Y-%m-%d-%Hh-%Mm-%Ss")))
-    fs::dir_create(docs_backup)
-    docs_backup <- absolute(docs_backup)
-    file.copy(from = file.path(s1$docs, "."), to = docs_backup,
-              recursive = TRUE, copy.date = TRUE)
-    on.exit(unlink(s1$docs, recursive = TRUE), add = TRUE)
-    on.exit(fs::dir_create(s1$docs), add = TRUE)
-    on.exit(file.copy(from = file.path(docs_backup, "."), to = s1$docs,
-                      recursive = TRUE, copy.date = TRUE), add = TRUE)
+    if (fs::dir_exists(s1$docs) && !dry_run) {
+      docs_backup <- tempfile(pattern = sprintf("docs-backup-%s-",
+                                                format(Sys.time(),
+                                                       "%Y-%m-%d-%Hh-%Mm-%Ss")))
+      fs::dir_create(docs_backup)
+      docs_backup <- absolute(docs_backup)
+      file.copy(from = file.path(s1$docs, "."), to = docs_backup,
+                recursive = TRUE, copy.date = TRUE)
+      on.exit(unlink(s1$docs, recursive = TRUE), add = TRUE)
+      on.exit(fs::dir_create(s1$docs), add = TRUE)
+      on.exit(file.copy(from = file.path(docs_backup, "."), to = s1$docs,
+                        recursive = TRUE, copy.date = TRUE), add = TRUE)
+    }
 
     step2 <- wflow_build(files = files_to_build, make = FALSE,
                          update = update, republish = republish,
